requirements.txt
numpy>=1.15
matplotlib>=3.2
brainpy-simulator>=1.0.0a0
